backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
tensorflow>=2.16.0
Pillow>=10.0.0
pefile==2023.2.7
